pulled in log manager changes. updated LogManager exception format. added __attribute__ to hlog call to tell compiler that the call is printf style. fixed ensuing warnings
git-svn-id: svn+ssh://svn/repos/scale-product/branches/overhaul_phase_1@8847 038091cf-3c17-4aaf-a619-8414e6e1260f

+++ ProcRunner.cpp
@@ -384,7 +384,9 @@ int ProcRunner::run(const FString& command,
             if (log_child) hlog(HLOG_DEBUG, "CHILD: %s", buf);
         }
 
-        if (i >= n) hlog(HLOG_DEBUG, "Child output logging truncated at %u lines", n);
+        if (i >= n) 
+            hlog(HLOG_DEBUG, 
+                 "Child output logging truncated at %llu lines", (u64) n);
 
         // cleanup
         in.close();
